refactor:LottoGenerator->Lotto로 이관
기존 테스트할 수 없는 영역을 Lotto 객체로 이관시켜킨 후,
테스트 환경에서는 Lotto생성자를 이용하여 진행
